rename DetailsScreen to DetailsMovieScreen
