+++ app/src/main/java/com/example/compmoviedb/presentation/navigation/SetupNavHost.kt
@@ -4,10 +4,9 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.example.compmoviedb.presentation.screens.details.DetailsScreen
+import com.example.compmoviedb.presentation.screens.detailsmovie.DetailsScreen
 import com.example.compmoviedb.presentation.screens.main.MainScreen
 import com.example.compmoviedb.presentation.screens.splash.SplashScreen
-import com.example.compmoviedb.presentation.utils.Constants
 
 @Composable
 fun SetupNavHost() {
@@ -27,7 +26,7 @@ fun SetupNavHost() {
                 navController = navController
             )
         }
-        composable(route = NavRoute.Details.route + "/{movieId}") { backStackEntry ->
+        composable(route = NavRoute.DetailsMovie.route + "/{movieId}") { backStackEntry ->
             DetailsScreen(
                 navController = navController,
                 movieId = backStackEntry.arguments?.getInt("movieId") ?: 0
