+++ app/src/main/java/com/example/compmoviedb/presentation/utils/Constants.kt
@@ -4,7 +4,7 @@ object Constants {
     object Screens {
         const val SPLASH_SCREEN = "splash_screen"
         const val MAIN_SCREEN = "main_screen"
-        const val DETAILS_SCREEN = "details_screen"
+        const val DETAILS_MOVIE_SCREEN = "details_movie_screen"
     }
 
     object Keys {
